# PERF: Removed the GIL from parts of the TextReader class