@jdeschenes 2015-10-09
I fixed a few issues with the build