@jdeschenes 2015-10-09
This was an issue with python3 and not limited to windows.