@jreback 2015-10-13
@jdeschenes 

- couple of questions for you above. 
- can you post the asv results here.
- pls add a note in the performance section (use this PR number)
- squash