@jdeschenes 2015-10-13
@jreback 

Where do I need to add information in the performance section? Is it in the what's new file in the documentation?