@jreback 2015-10-13
whatsnew/v0.17.1 (Performance section)