@jreback 2015-10-25
@jdeschenes looks good. can you

- add a note in whatsnew
- rebase