@mrocklin 2015-11-03
What's the status on this @jdeschenes?  I'd like to include this work in a talk happening tomorrow.  It'd be awesome to be able to say that this was in master rather than in a branch.