@jdeschenes 2015-11-03
I will get the final changes tomorrow.