@jdeschenes 2015-11-04
@jreback, The changes have been implemented. Let me know if there is anything else that needs to be done.